First cut at adding account creation endpoints.
Still quite a bit missing -- especially persisting users in the
database.  Currently the registration fails because the dummy user
created is different from start to finish endpoints. We should also
set very short cookie lifetimes (just the expected time to complete
the account creation process. We probably don't need to use the same
key every time for signing cookies but could persist it if we wanted.